+++ instrumentation/sinks/os/exec/orchestrion.yml
@@ -19,7 +19,7 @@ aspects:
     advice:
       - prepend-statements:
           imports:
-            zentransits: github.com/AikidoSec/firewall-go/internal/vulnerabilities/transits
+            zentransits: github.com/AikidoSec/firewall-go/instrumentation/transits
           template: |-
             if examineCommand := zentransits.ExamineCommand(); examineCommand != nil {
               _aikido_err := examineCommand({{ .Function.Receiver}}.ctx, "exec.Cmd.{{ .Function.Name }}", {{ .Function.Receiver }}.Args)
