+++ internal/examine/command_test.go
@@ -1,4 +1,4 @@
-package internal
+package examine
 
 import (
 	"context"
@@ -43,37 +43,6 @@ func TestExamineCommand_Disabled(t *testing.T) {
 	err := examineCommand(ctx, "os/exec.Command", maliciousArgs)
 
 	require.NoError(t, err, "Should not detect shell injection when zen is disabled")
-
-	select {
-	case <-mockClient.AttackDetectedEventSent:
-		t.Fatal("No attack should be detected when zen is disabled")
-	case <-time.After(50 * time.Millisecond):
-		// Expected: no attack detected
-	}
-}
-
-func TestExaminePath_Disabled(t *testing.T) {
-	originalDisabled := config.IsZenDisabled()
-	defer config.SetZenDisabled(originalDisabled)
-
-	originalClient := agent.GetCloudClient()
-	originalBlocking := config.IsBlockingEnabled()
-	config.SetBlocking(true)
-	defer func() {
-		config.SetBlocking(originalBlocking)
-		agent.SetCloudClient(originalClient)
-	}()
-
-	mockClient := testutil.NewMockCloudClient()
-	agent.SetCloudClient(mockClient)
-
-	config.SetZenDisabled(true)
-
-	maliciousPath := "../../etc/passwd"
-
-	err := examinePath("os.OpenFeil", []string{maliciousPath}, false)
-
-	require.NoError(t, err, "Examine should return early with no error when zen is disabled")
 
 	select {
 	case <-mockClient.AttackDetectedEventSent:
