+++ internal/examine/transits.go
@@ -1,0 +1,12 @@
+package examine
+
+import (
+	"github.com/AikidoSec/firewall-go/instrumentation/transits"
+)
+
+func SetupTransits() {
+	transits.SetTransits(
+		examinePath,
+		examineCommand,
+	)
+}
